Create Project 1 - Differences and Trends
The purpose of this project is to identify and visualize differences and trends in the market values of housing units over a specific period. The data: Housing Affordability Data System of the U.S Department of Housing and Urban Development [HADS]. HADS is a set of housing unit-level data sets that aims to provide consistent measures of affordability and cost burdens over a long period.